# docs/openapi.yaml
openapi: 3.0.2
servers:
  - url: https://api.mainnet.hiro.so/
    description: Mainnet
  - url: https://api.nakamoto.testnet.hiro.so/
    description: Nakamoto Testnet
  - url: https://api.testnet.hiro.so/
    description: Testnet
  - url: http://localhost:3999/
    description: Local
info:
  title: Stacks Blockchain API
  version: 'STACKS_API_VERSION'
  description: |
    Welcome to the API reference overview for the [Stacks Blockchain API](https://docs.hiro.so/stacks-blockchain-api).

    [Download Postman collection](https://hirosystems.github.io/stacks-blockchain-api/collection.json)
...
paths:
...
  /extended/v1/tokens/nft/holdings:
    get:
      operationId: get_nft_holdings
      summary: Non-Fungible Token holdings
      description: |
        Retrieves the list of Non-Fungible Tokens owned by the given principal (STX address or Smart Contract ID).
        Results can be filtered by one or more asset identifiers and can include metadata about the transaction that made the principal own each token.

        More information on Non-Fungible Tokens on the Stacks blockchain can be found [here](https://docs.stacks.co/write-smart-contracts/tokens#non-fungible-tokens-nfts).
      tags:
        - Non-Fungible Tokens
      parameters:
        - name: principal
          in: query
          description: token owner's STX address or Smart Contract ID
          required: true
          schema:
            type: string
            example: "SPNWZ5V2TPWGQGVDR6T7B6RQ4XMGZ4PXTEE0VQ0S.marketplace-v3"
        - name: asset_identifiers
          in: query
          description: identifiers of the token asset classes to filter for
          required: false
          style: form
          explode: true
          schema:
            type: array
            example: "SPQZF23W7SEYBFG5JQ496NMY0G7379SRYEDREMSV.Candy::candy"
            items:
              type: string
        - name: limit
          in: query
          description: max number of tokens to fetch
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: index of first tokens to fetch
          required: false
          schema:
            type: integer
            default: 0
            example: 42000
        - name: unanchored
          in: query
          description: whether or not to include tokens from unconfirmed transactions
          required: false
          schema:
            type: boolean
            example: true
            default: false
        - name: tx_metadata
          in: query
          description: whether or not to include the complete transaction metadata instead of just `tx_id`. Enabling this option can affect performance and response times.
          required: false
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: List of Non-Fungible Token holdings
          content:
            application/json:
              schema:
                $ref: ./api/tokens/get-non-fungible-token-holdings.schema.json
              examples:
                default:
                  value:
                    $ref: ./api/tokens/get-non-fungible-token-holdings.example.schema.json
                with transaction metadata:
                  value:
                    $ref: ./api/tokens/get-non-fungible-token-holdings-tx-metadata.example.schema.json
  /extended/v1/tokens/nft/history:
    get:
      operationId: get_nft_history
      summary: Non-Fungible Token history
      description: |
        Retrieves all events relevant to a Non-Fungible Token. Useful to determine the ownership history of a particular asset.

        More information on Non-Fungible Tokens on the Stacks blockchain can be found [here](https://docs.stacks.co/write-smart-contracts/tokens#non-fungible-tokens-nfts).
      tags:
        - Non-Fungible Tokens
      parameters:
        - name: asset_identifier
          in: query
          description: token asset class identifier
          required: true
          schema:
            type: string
            example: "SP2X0TZ59D5SZ8ACQ6YMCHHNR2ZN51Z32E2CJ173.the-explorer-guild::The-Explorer-Guild"
        - name: value
          in: query
          description: hex representation of the token's unique value
          required: true
          schema:
            type: string
            example: "0x0100000000000000000000000000000803"
        - name: limit
          in: query
          description: max number of events to fetch
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: index of first event to fetch
          required: false
          schema:
            type: integer
            default: 0
            example: 42000
        - name: unanchored
          in: query
          description: whether or not to include events from unconfirmed transactions
          required: false
          schema:
            type: boolean
            default: false
            example: true
        - name: tx_metadata
          in: query
          description: whether or not to include the complete transaction metadata instead of just `tx_id`. Enabling this option can affect performance and response times.
          required: false
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: Non-Fungible Token event history
          content:
            application/json:
              schema:
                $ref: ./api/tokens/get-non-fungible-token-history.schema.json
              examples:
                default:
                  value:
                    $ref: ./api/tokens/get-non-fungible-token-history.example.schema.json
                with transaction metadata:
                  value:
                    $ref: ./api/tokens/get-non-fungible-token-history-tx-metadata.example.schema.json
  /extended/v1/tokens/nft/mints:
    get:
      operationId: get_nft_mints
      summary: Non-Fungible Token mints
      description: |
        Retrieves all mint events for a Non-Fungible Token asset class. Useful to determine which NFTs of a particular collection have been claimed.

        More information on Non-Fungible Tokens on the Stacks blockchain can be found [here](https://docs.stacks.co/write-smart-contracts/tokens#non-fungible-tokens-nfts).
      tags:
        - Non-Fungible Tokens
      parameters:
        - name: asset_identifier
          in: query
          description: token asset class identifier
          required: true
          schema:
            type: string
            example: "SP2X0TZ59D5SZ8ACQ6YMCHHNR2ZN51Z32E2CJ173.the-explorer-guild::The-Explorer-Guild"
        - name: limit
          in: query
          description: max number of events to fetch
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: index of first event to fetch
          required: false
          schema:
            type: integer
            default: 0
            example: 42000
        - name: unanchored
          in: query
          description: whether or not to include events from unconfirmed transactions
          required: false
          schema:
            type: boolean
            example: true
            default: false
        - name: tx_metadata
          in: query
          description: whether or not to include the complete transaction metadata instead of just `tx_id`. Enabling this option can affect performance and response times.
          required: false
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: Non-Fungible Token mints
          content:
            application/json:
              schema:
                $ref: ./api/tokens/get-non-fungible-token-mints.schema.json
              examples:
                default:
                  value:
                    $ref: ./api/tokens/get-non-fungible-token-mints.example.schema.json
                with transaction metadata:
                  value:
                    $ref: ./api/tokens/get-non-fungible-token-mints-tx-metadata.example.schema.json
...